Candidato Sergio Mateus

+++ src/main/java/br/com/alelo/consumer/consumerpat/respository/ConsumerRepository.java
@@ -6,17 +6,17 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ConsumerRepository extends JpaRepository<Consumer, Integer> {
+public interface ConsumerRepository extends JpaRepository<Consumer, Long> {
 
     @Query(nativeQuery = true, value = "select * from Consumer")
     List<Consumer> getAllConsumersList();
 
     @Query(nativeQuery = true, value = "select * from Consumer where FOOD_CARD_NUMBER = ? ")
-    Consumer findByFoodCardNumber(int cardNumber);
+    Consumer findByFoodCardNumber(Long cardNumber);
 
     @Query(nativeQuery = true, value = "select * from Consumer where FUEL_CARD_NUMBER = ? ")
-    Consumer findByFuelCardNumber(int cardNumber);
+    Consumer findByFuelCardNumber(Long cardNumber);
 
     @Query(nativeQuery = true, value = "select * from Consumer where DRUGSTORE_NUMBER = ? ")
-    Consumer findByDrugstoreNumber(int cardNumber);
+    Consumer findByDrugstoreNumber(Long cardNumber);
 }
